+++ tests/utils.test.mjs
@@ -20,6 +20,7 @@ import {
   max,
   median,
   min,
+  once,
   round,
   secureRandom,
   sleep,
@@ -239,5 +240,20 @@ Deno.test('Utils test suite', async (t) => {
     expect(max(1, 2)).toBe(2)
     expect(max(2, 1)).toBe(2)
     expect(max(1, 1)).toBe(1)
+  })
+
+  await t.step('Verify once()', () => {
+    let called = 0
+    const fn = () => ++called
+    const onceFn = once(fn, this)
+    const result1 = onceFn()
+    expect(called).toBe(1)
+    expect(result1).toBe(1)
+    const result2 = onceFn()
+    expect(called).toBe(1)
+    expect(result2).toBe(1)
+    const result3 = onceFn()
+    expect(called).toBe(1)
+    expect(result3).toBe(1)
   })
 })