+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure worker ready response can be received only once.
+
+<!-- - Ensure pool ready event can be emitted only once. -->
+
 ## [0.0.6] - 2023-10-20
 
 ### Added